Extract and export the `tuple` serializer
