+++ packages/umi-serializer-beet/src/index.ts
@@ -1,5 +1,6 @@
 export * from './createBeetSerializer';
 export * from './bytes';
 export * from './errors';
+export * from './tuple';
 export * from './plugin';
 export * from './pubkey';
