+++ packages/umi-serializer-data-view/src/index.ts
@@ -1,5 +1,6 @@
 export * from './createDataViewSerializer';
 export * from './bytes';
 export * from './errors';
+export * from './tuple';
 export * from './plugin';
 export * from './pubkey';
